+++ intelmq/bots/experts/maxmind_geoip/expert.py
@@ -57,23 +57,23 @@ class GeoIPExpertBot(Bot):
                 if self.registered:
                     if info.registered_country.iso_code:
                         event.add(geo_key % "cc", info.registered_country.iso_code,
-                                  overwrite=self.parameters)
+                                  overwrite=self.parameters.overwrite)
                 else:
                     if info.country.iso_code:
                         event.add(geo_key % "cc", info.country.iso_code,
-                                  overwrite=self.parameters)
+                                  overwrite=self.parameters.overwrite)
 
                 if info.location.latitude:
                     event.add(geo_key % "latitude", info.location.latitude,
-                              overwrite=self.parameters)
+                              overwrite=self.parameters.overwrite)
 
                 if info.location.longitude:
                     event.add(geo_key % "longitude", info.location.longitude,
-                              overwrite=self.parameters)
+                              overwrite=self.parameters.overwrite)
 
                 if info.city.name:
                     event.add(geo_key % "city", info.city.name,
-                              overwrite=self.parameters)
+                              overwrite=self.parameters.overwrite)
 
             except geoip2.errors.AddressNotFoundError:
                 pass