+++ CHANGELOG.md
@@ -87,6 +87,8 @@ IntelMQ no longer supports Python 3.5 (and thus Debian 9 and Ubuntu 16.04), the 
 - Added `intelmq.bots.experts.splunk_saved_search.expert` (PR#1666 by Karl-Johan Karlsson).
 - `intelmq.bots.experts.sieve.expert`:
   - Added possibility to give multiple queue names for the `path` directive (#1462).
+- `intelmq.bots.experts.maxmind_geoip.expert`:
+  - Fixed handing over of `overwrite` parameter to `event.add`
 
 #### Outputs
 - `intelmq.bots.outputs.rt`: Added Request Tracker output bot (PR#1589 by Marius Urkis).